+++ src/molecules/footer-menu/FooterMenu.vue
@@ -1,0 +1,27 @@
+<template src="./FooterMenu.html" />
+
+<script>
+import AIcon from '../../atoms/icon/Icon.vue'
+import ALink from '../../atoms/link/Link.vue'
+import AButton from '../../atoms/button/Button.vue'
+import AList from '../../atoms/list/List.vue'
+import AListItem from '../../atoms/list-item/ListItem.vue'
+import ACollapse from '../../molecules/collapse/Collapse.vue'
+
+import AFooter from './FooterMenu.js'
+
+export default {
+  name: 'AlpacaFooter',
+  components: {
+    AButton,
+    ALink,
+    AIcon,
+    AList,
+    AListItem,
+    ACollapse
+  },
+  mixins: [AFooter]
+}
+</script>
+
+<style lang="scss" src="./FooterMenu.scss" />