+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for Preformatted component
 - More tests in input component
 - Tests for Textarea component
+- FooterMenu component
 - Tests for Select component
 - Custom icon in Select component
 
@@ -39,6 +40,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Price component (removed props, changed classes, removed useless storeies)
 - Collapse style
 - Move modal to atomic design, update template and component js logic
+- Refactor Newsletter component
+- Refactor Collapse component
+- Refactor Footer component
 - Move breadcrumbs to atomic design, update styles and template
 - Remove 'selected' story in Selected coomponents
 - Refactor styles in Select component